// src/generated/api-types.ts
export interface operations {
  "Authorize Application": {
// ...
    responses: {
      /** @description OK */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content?: never;
      };
    };
  };
// ...
  "Start watching": {
// ...
    responses: {
      /** @description Created */
      201: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json":
            | {
                id?: number;
                action?: string;
                progress?: number;
                movie?: {
                  title?: string;
                  year?: number;
                  anime_type?: unknown;
                  ids?: {
                    simkl?: number;
                    slug?: string;
                    imdb?: string;
                    letterslug?: string;
                    tmdb?: string;
                    tvdbslug?: string;
                  };
                };
              }
            | {
                id?: number;
                action?: string;
                progress?: number;
                show?: {
                  title?: string;
                  year?: number;
                  ids?: {
                    simkl?: number;
                    imdb?: string;
                  };
                };
                episode?: {
                  season?: number;
                  number?: number;
                  title?: string;
                };
              }
            | {
                id?: number;
                action?: string;
                progress?: number;
                episode?: {
                  season?: number;
                  number?: number;
                  title?: string;
                  tvdb_season?: number;
                  tvdb_number?: number;
                };
                anime?: {
                  title?: string;
                  year?: number;
                  anime_type?: string;
                  ids?: {
                    simkl?: number;
                    slug?: string;
                    mal?: string;
                    anilist?: string;
                    kitsu?: string;
                    anidb?: string;
                  };
                };
              }
            | {
                id?: number;
                action?: string;
                progress?: number;
                anime?: {
                  title?: string;
                  year?: number;
                  anime_type?: string;
                  ids?: {
                    simkl?: number;
                    slug?: string;
                    mal?: string;
                    anilist?: string;
                    kitsu?: string;
                    anidb?: string;
                  };
                };
              };
        };
      };
    };
  };
// ...
}